refactor: update time formatting and message assertion logic for consistency and clarity
