+++ packages/logger-plugin-console/src/helpers/helper-assert-message.ts
@@ -2,14 +2,13 @@ import type { LoggerMessageObject } from '@hyperse/logger';
 
 export const assertMessage = <T extends LoggerMessageObject>(
   message: T | string
-): T => {
+): LoggerMessageObject => {
   if (typeof message === 'string') {
-    const newMessage = {
+    return {
       message: message,
       name: undefined,
       stack: undefined,
     };
-    return newMessage as T;
   }
   return message;
 };
