+++ packages/logger-plugin-std/src/helpers/helper-merge-options.ts
